fix: Address Copilot code review feedback
- Remove duplicate AnalyticsEvent type definition, import from db.ts
- Move AnalyticsEventType to db.ts for single source of truth
- Use crypto.randomUUID() for secure session ID generation
- Make Lingui translations reactive with useMemo and locale dependency
- Update translations when locale changes via useEffect

Resolves all Copilot and GitHub Security review comments.
